norm

+++ includes/ft_ping.h
@@ -63,7 +63,7 @@ int				dns_err(t_ping_data *data, int err, struct addrinfo *hints, \
 t_ping_pkt		*build_pkt(int msg_count);
 struct msghdr	*build_msg(struct sockaddr *addr_struct);
 void			sig_handler(int numSig);
-void			save_stats(t_ping_data *data, long int *delay);
+void			save_stats(t_ping_data *data, int *delay);
 void			count_success(t_ping_data *data);
 void			print_stats(t_ping_data *data, int msg_count, int delay);
 void			print_pkt_stats(t_ping_data *data, int received_size, \
